Make the top-level API more object-oriented.
Interface is now a method of *Object because it does not rely on the
connection.

The names of *Connection methods have been shortened.
    CallMethod => Call
    EmitSignal => Emit
    GetObject => Object
    AddSignalHandler => Handle

The Connection.Call method now takes a *Method and a variadic argument
list. Similarly, Connection.Emit takes a *Signal and a variadic argument
list.

*Method and *Signal objects are created by the Method and Signal methods
of *Interface objects. Each method can return an error if the given name
is invalid.